#1 Add deployment, pivot config fields to Go app, remove dead code
The ces-importer will run on the target system in two parts:
1. the main application which controls the timing of the sync jobs
2. the sync jobs themselves

The main application will start the jobs by golang API means rather than
by Helm chart YAMLs. This means the app config as well as the actual
Helm chart must be restructured to be fit for customization.

Thus, this commit restructures the existing configuration values.yaml
and applies them to a configmap. This configmap will be mounted in the
main deployment by now. What it takes to create the sync job is up to
the implementor of the sync job f

This commit also removes superfluous fields that will be returned by
calling the exporter API which will return data required by the rsync
copy process.