#4 stop and start dogus
This commit adds the necessary k8s support to the ces-importer to stop
and start dogus before and after an imaginary (not fully working) sync.

To do so the deployments must be scaled down and up respectively. Since
dogus do not support a replica count >=2, this commit takes the freedom
to simply use replica values of zero and one for down and up scaling.

As of now, this is a complexity problem to save this information. If
this is an issue, the original replica count must be saved and used to
restore the original value.

To organize the new k8s functionality all API related code was
reorganized into exporter and importer API packages. Also the usual
bells and whistles are needed: kube config, k8s API client etc. so the
main go file is heavily refactored to supply all the clients. The k8s
namespace is taken from the Helm chart but under a different environment
variable to avoid ambiguities with future namespaces regarding the
exporter.