#4 Add cron tests, remove somewhat package stuttering, move validation
cron.NewPackageLooper() is a bit of a golang stutter when cron.New()
totally suffices to reflect on creating a timely function executor.

Also, this commit moves the cron-like expression validation to the
constructor call time instead of the expression execution time. This,
because catching errors sooner is better and costs less debugging time.