#4 Replace context Warn logging with simple Warn()
In terms of slog, the context may or may not provide informational
tidbits -- which we don't use here when we probably should. But that
would take a bit more logging infrastructure like so which would blow up
the codebase unnecessarily (by now):

```go
type ContextHandler struct {
    slog.Handler
}
// Handle overrides the default Handle method to add context values.
func (h *ContextHandler) Handle(ctx context.Context, r slog.Record) error {
    if requestID, ok := ctx.Value("request_id").(string); ok {
        r.AddAttrs(slog.String("request_id", requestID))
    }
    if userID, ok := ctx.Value("user_id").(string); ok {
        r.AddAttrs(slog.String("user_id", userID))
    }
    return h.Handler.Handle(ctx, r)
}
```
and so on.