[#66] Fix misformat mail header
The string formatting for the from header already contained a CRLF. Due to non-deterministic iteration over the header map there has been an empty line between headers, which leads to the error in the format of the email.

For better comparison a deterministic iteration over the headers has been introduced.