[#66] Use io.Copy instead of bufio.Scanner to stream logs from import-job to coordinator
bufio.Scanner uses a buffer to marshal logs and send them to the log of the coordinator. In case of very long log lines this leads to "too long token" errors. Instead of increasing the buffer size, io.Copy is used to stream the logs directly to the log writer, which avoids marshalling and also improves the performance.